Exec command: Order and repetition of jobs #15
