+++ src/main/java/io/bootique/job/scheduler/DefaultScheduler.java
@@ -9,8 +9,8 @@ import io.bootique.job.runnable.JobFuture;
 import io.bootique.job.runnable.JobResult;
 import io.bootique.job.runnable.RunnableJob;
 import io.bootique.job.runnable.RunnableJobFactory;
-import io.bootique.job.scheduler.execution.Execution;
 import io.bootique.job.scheduler.execution.ExecutionFactory;
+import io.bootique.job.scheduler.execution.Execution;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.TaskScheduler;
