+++ src/main/java/io/bootique/job/scheduler/execution/ExecutionVisitor.java
@@ -1,0 +1,8 @@
+package io.bootique.job.scheduler.execution;
+
+import java.util.Set;
+
+public interface ExecutionVisitor {
+
+    void visitExecutionStep(Set<JobExecution> jobExecutions);
+}
